online_inference/requirements.txt
jedi==0.18.1
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.0
marshmallow==3.18.0
marshmallow-dataclass==8.5.9
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas~=1.5.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyparsing==3.0.9
pyrsistent==0.19.1
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML~=5.4.1
scikit-learn~=1.1.2
scipy==1.9.3
typing-extensions==4.4.0
typing-inspect==0.8.0
sklearn~=0.0
click~=8.1.3
fastapi~=0.85.0
pydantic~=1.10.2